// web/src/types/generated/types.gen.ts
export type CourseOutput = {
    /**
     * Uid
     *
     * UID
     */
    uid: string;
    /**
     * Title
     *
     * Course title
     */
    title: string;
    /**
     * Description
     *
     * Course description
     */
    description?: string | null;
    /**
     * Order
     *
     * Display order
     */
    order: number;
    /**
     * Is Archived
     *
     * Archive flag
     */
    is_archived: boolean;
    /**
     * Created Date
     *
     * Created date
     */
    created_date: string;
    /**
     * Updated Date
     *
     * Updated date
     */
    updated_date: string;
    /**
     * Directions
     *
     * List of course directions
     */
    directions?: Array<DirectionFull>;
};
// ...
export type DirectionFull = {
    /**
     * Uid
     *
     * UID
     */
    uid: string;
    /**
     * Title
     *
     * Direction title
     */
    title: string;
    /**
     * Description
     *
     * Direction description
     */
    description?: string | null;
    /**
     * Is Archived
     *
     * Archive flag
     */
    is_archived: boolean;
    /**
     * Created Date
     *
     * Created date
     */
    created_date: string;
    /**
     * Updated Date
     *
     * Updated date
     */
    updated_date: string;
};
// ...
export type RoleFull = {
    /**
     * Uid
     *
     * Role UID
     */
    uid: string;
    /**
     * Title
     *
     * Role title (fe. teacher, admin, manager, etc.)
     */
    title: 'student' | 'teacher' | 'admin' | 'manager' | 'promoter' | 'salesman';
};
// ...
export type StudentThemeInput = {
    /**
     * Student Uid
     *
     * Student UID
     */
    student_uid: string;
    /**
     * Theme Uid
     *
     * Theme UID
     */
    theme_uid: string;
    /**
     * Comment
     *
     * Comment
     */
    comment?: string | null;
    /**
     * Is Finished
     *
     * Is finished
     */
    is_finished?: boolean;
};

/**
 * student.StudentThemeListedOutput
 */
export type StudentThemeListedOutput = {
    /**
     * Items
     *
     * List of student theme progress
     */
    items: Array<StudentThemeOutput>;
    /**
     * Count
     *
     * Total count
     */
    count: number;
};

/**
 * student.StudentThemeOutput
 */
export type StudentThemeOutput = {
    /**
     * Uid
     *
     * UID
     */
    uid: string;
    /**
     * Student Uid
     *
     * Student UID
     */
    student_uid: string;
    /**
     * Theme Uid
     *
     * Theme UID
     */
    theme_uid: string;
    /**
     * Comment
     *
     * Comment
     */
    comment?: string | null;
    /**
     * Is Finished
     *
     * Is finished
     */
    is_finished: boolean;
    /**
     * Created Date
     *
     * Created date
     */
    created_date: string;
    /**
     * Updated Date
     *
     * Updated date
     */
    updated_date: string;
    /**
     * User data
     */
    user: UserOutput;
    /**
     * Theme data
     */
    theme: ThemeOutput;
};
// ...
export type ThemeOutput = {
    /**
     * Uid
     *
     * UID
     */
    uid: string;
    /**
     * Title
     *
     * Theme title
     */
    title: string;
    /**
     * Description
     *
     * Theme description
     */
    description?: string | null;
    /**
     * Course Uid
     *
     * Parent course UID
     */
    course_uid: string;
    /**
     * Order
     *
     * Display order
     */
    order: number;
    /**
     * Is Archived
     *
     * Archive flag
     */
    is_archived: boolean;
    /**
     * Created Date
     *
     * Created date
     */
    created_date: string;
    /**
     * Updated Date
     *
     * Updated date
     */
    updated_date: string;
};
// ...
export type UserOutput = {
    /**
     * Uid
     *
     * UID
     */
    uid: string;
    /**
     * Fullname
     *
     * Fullname
     */
    fullname: string;
    /**
     * Email
     *
     * Email
     */
    email?: string | null;
    /**
     * Registration Date
     *
     * Registration datetime
     */
    registration_date: string;
    /**
     * Last Login Date
     *
     * Last login datetime
     */
    last_login_date?: string | null;
    /**
     * Last Update Date
     *
     * Datetime when user info was updated
     */
    last_update_date: string;
    /**
     * Is Archived
     *
     * Archive user or not.
     */
    is_archived: boolean;
    /**
     * Roles
     *
     * List of user roles
     */
    roles: Array<RoleFull>;
    /**
     * Student Uid
     *
     * UID of the student linked to this user
     */
    student_uid?: string | null;
};